pyproject.toml
[project]
name = "hca-atlas-tree"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.66.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
]
